refactor: remove extension module
It will soon be replaced with a much better extension API interface.

BREAKING CHANGE: Importers must import definitions from their actual definition locations. The extension module no longer exists.
