--- src/extension/index.ts
@@ -1,6 +1,0 @@
-/**
- * This module exports the public API used by CXP extensions.
- */
-
-export { Connection } from '../server/server'
-export * from '../protocol'
